Revert "Change model"
This reverts commit 382ffd76502e38db81816e4519aa723e2885ae42.

+++ src/main/java/JavaWeb/GameAccount/repositories/ICategoryRepository.java
@@ -1,16 +1,11 @@
 package JavaWeb.GameAccount.repositories;
 
 import JavaWeb.GameAccount.model.Category;
-import JavaWeb.GameAccount.model.Product;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
 @Repository
 
-public interface ICategoryRepository extends JpaRepository<Category, Long>{
+public interface ICategoryRepository extends JpaRepository<Category, Integer>{
     Category findByLink(String link);
+
 }
